+++ src/Components/index.ts
@@ -14,3 +14,4 @@ export * from './ImageSlider/ImageSlider';
 export * from './DateTimePicker/DateTimePicker';
 export * from './RangeSlider/RangeSlider';
 export * from './AuctionEditForm/AuctionEditForm';
+export * from './ProductRegisterForm/ProductRegisterForm';
