+++ src/Pages/ProductRegisterPage/ProductRegisterPage.type.ts
@@ -2,9 +2,6 @@ export type ProductRegisterForm = {
   images: FileList;
   title: string;
   content: string;
-  startAt: string;
-  startPrice: string;
-  instant: boolean;
   buyNowPrice: string;
   isAuction: boolean;
   auctionStartPrice: string;
